refactor(dpp): give the ranked prefix-overlap rule its own file
validate_no_ranked_prefix_overlap moves from v3/mod.rs into
v3/ranked_prefix_overlap.rs — one frozen unit of generation-3 grammar
per file, same layout rationale as the mode-detection version modules.
No behavior change.
